Exploding enemies added

+++ _blok/Assets/Scripts/Entities/BBIDamageable.cs
@@ -2,6 +2,6 @@
 using System.Collections;
 
 public interface BBIDamageable {
-	void TakeHit(float power, Collider collider);
-	void TakeHit(float power, Collider collider, BBKnockback knockback);
+	void TakeHit(int power, Collider collider);
+	void TakeHit(int power, Collider collider, BBKnockback knockback);
 }
