+++ _blok/Assets/Scripts/Widgits/Constants/BBEntityConstants.cs
@@ -4,6 +4,6 @@ using System.Collections;
 //This class contains constant values entity properties (e.g. default knockback)
 public class BBEntityConstants {
 	public const float defaultKnockbackMagnitude = 10.0f;
-	public const float defaultKnockbackDamage = 2.0f;
+	public const int defaultKnockbackDamage = 2;
 	public const float defaultKnockbackTime = .2f;
 }
